requirements.txt
streamlit
Pillow
imagehash
numpy